+++ src/modules/cloudinary/cloudinary.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { ConfigModule } from '@nestjs/config';
+
+import { CloudinaryService } from './cloudinary.service';
+
+@Module({
+  imports: [ConfigModule],
+  providers: [CloudinaryService],
+  exports: [CloudinaryService],
+})
+export class CloudinaryModule {}
