+++ src/modules/gpt/constants/index.ts
@@ -17,16 +17,3 @@ export enum ModelSpeech {
 export enum ModelImage {
   DALL_E_3 = 'dall-e-3',
 }
-
-export const IMAGE_SIZE_WIDTH_DEFAULT = 1024;
-export const IMAGE_SIZE_HEIGHT_DEFAULT = 1024;
-
-export const SBER_OAUTH = 'https://ngw.devices.sberbank.ru:9443/api/v2/oauth';
-
-export const GIGA_CHAT = 'https://gigachat.devices.sberbank.ru/api/v1';
-export const GIGACHAT_API_PERS = 'GIGACHAT_API_PERS';
-export const GIGA_CHAT_ACCESS_TOKEN = 'giga-chat-access-token';
-
-export const SMART_SPEEECH = 'https://smartspeech.sber.ru/rest/v1';
-export const SALUTE_SPEECH_PERS = 'SALUTE_SPEECH_PERS';
-export const SALUTE_SPEECH_ACCESS_TOKEN = 'salute-speech-access-token';
