+++ src/modules/sber/constants/sber.constants.ts
@@ -1,0 +1,9 @@
+export const SBER_OAUTH = 'https://ngw.devices.sberbank.ru:9443/api/v2/oauth';
+
+export const GIGA_CHAT = 'https://gigachat.devices.sberbank.ru/api/v1';
+export const GIGACHAT_API_PERS = 'GIGACHAT_API_PERS';
+export const GIGA_CHAT_ACCESS_TOKEN = 'giga-chat-access-token';
+
+export const SMART_SPEEECH = 'https://smartspeech.sber.ru/rest/v1';
+export const SALUTE_SPEECH_PERS = 'SALUTE_SPEECH_PERS';
+export const SALUTE_SPEECH_ACCESS_TOKEN = 'salute-speech-access-token';
