+++ src/modules/sber/sber.module.ts
@@ -1,0 +1,22 @@
+import { HttpModule } from '@nestjs/axios';
+import { Module } from '@nestjs/common';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+
+import { SberService } from './sber.service';
+
+@Module({
+  imports: [
+    HttpModule.registerAsync({
+      imports: [ConfigModule],
+      inject: [ConfigService],
+      useFactory: async (configService: ConfigService) => ({
+        timeout: configService.get('http.timeout'),
+        maxRedirects: configService.get('http.maxRedirects'),
+      }),
+    }),
+    ConfigModule,
+  ],
+  providers: [SberService],
+  exports: [SberService],
+})
+export class SberModule {}
